lambda/main/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
argcomplete==3.1.2
async-timeout==4.0.3
attrs==23.1.0
blis==0.7.10
boto3==1.28.47
botocore==1.31.47
Brotli==1.0.9
brotlipy==0.7.0
cachetools==5.3.0
catalogue==2.0.9
certifi==2023.7.22
click==8.1.3
confection==0.1.3
cymem==2.0.7
decorator==4.4.2
docopt==0.6.2
en-core-web-trf @ https://github.com/explosion/spacy-models/releases/download/en_core_web_trf-3.6.1/en_core_web_trf-3.6.1-py3-none-any.whl#sha256=f9a1f0ae83954c49ebf92c0f7043ea02de51db59bf6cd3ed849d46928612c806
exceptiongroup==1.1.2
ffmpeg==1.4
filelock==3.12.4
Flask==2.2.3
frozenlist==1.4.0
fsspec==2023.9.0
grpcio==1.51.3
grpcio-status==1.51.3
gunicorn==21.2.0
h11==0.14.0
httplib2==0.21.0
huggingface-hub==0.17.1
imageio==2.26.0
imageio-ffmpeg==0.4.8
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpointer==2.0
langcodes==3.3.0
MarkupSafe==2.1.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.4
murmurhash==1.0.9
mutagen==1.46.0
networkx==3.1
numpy==1.24.2
openai==0.28.0
outcome==1.2.0
pathy==0.10.2
Pillow==9.4.0
pipreqs==0.4.13
pipx==1.2.0
preshed==3.0.8
proglog==0.1.10
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.17
pydantic==2.3.0
pydantic_core==2.6.3
pydub==0.25.1
pyparsing==3.0.9
PyPDF2==3.0.1
python-dotenv==1.0.0
regex==2023.8.8
rsa==4.9
s3transfer==0.6.2
safetensors==0.3.3
selenium==4.11.2
smart-open==6.4.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.6.1
spacy-alignments==0.9.0
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.2.5
SpeechRecognition==3.10.0
srsly==2.4.7
sympy==1.12
tokenizers==0.13.3
transformers==4.30.2
trio==0.22.2
trio-websocket==0.10.3
typer==0.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
userpath==1.9.1
wasabi==1.1.2
websockets==10.4
Werkzeug==2.2.3
wsproto==1.2.0
yarg==0.1.9
yarl==1.9.2